Qashbits-main/requirements.txt
requests
bs4
ini-parser
pillow
pytesseract
